src/functions/bd_update_sales/requirements.txt
retry
python-dotenv
psychopg2-binary